X86 32bit port

+++ userspace/include/signal.h
@@ -63,14 +63,20 @@ typedef struct {
     int si_code;
 } siginfo_t;
 
-typedef unsigned long sigset_t;
+/* 64-bit to match the kernel ABI (struct sigaction uses u64 sa_flags/sa_mask,
+ * sigset_t is a u64 bitmask). `unsigned long` is 8 bytes on x86_64 but only 4 on
+ * the 32-bit port, which shifted sa_restorer/sa_mask to the wrong offsets — the
+ * kernel then read a garbage sa_restorer and killed any process that installed a
+ * real signal handler (M15). Use a fixed 64-bit type so the layout matches on
+ * both architectures. */
+typedef unsigned long long sigset_t;
 
 struct sigaction {
     union {
         sighandler_t sa_handler;
         void (*sa_sigaction)(int, siginfo_t *, void *);
     };
-    unsigned long sa_flags;
+    unsigned long long sa_flags;
     void (*sa_restorer)(void);
     sigset_t sa_mask;
 };
